Criado serviço de excluir e largar entrega e feito ajustes

+++ src/main/java/br/com/eubusco/server/dao/GenericDAO.java
@@ -67,7 +67,7 @@ public interface GenericDAO<T> {
 	 * @return
 	 * @throws BusinessServerException
 	 */
-	public abstract T excluirPorId(Long id);
+	public abstract T excluirPorId(Integer id);
 
 	/**
 	 * Busca pelo atributo Id da classe informada
